esyshape/requirements.txt
alabaster==0.7.12
arrow==0.15.4
attrs==19.3.0
Babel==2.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
descartes==1.1.0
docutils==0.15.2
Fiona==1.8.11
geopandas==0.4.1
idna==2.8
imagesize==1.1.0
Jinja2==2.10.3
jinja2-time==0.2.0
kiwisolver==1.1.0
make==0.1.5
MarkupSafe==1.1.1
matplotlib==3.0.3
munch==2.5.0
natsort==6.0.0
numpy==1.16.2
numpydoc==0.8.0
packaging==19.2
pandas==0.24.2
pathlib==1.0.1
psycopg2==2.8.4
Pygments==2.5.2
pyparsing==2.4.5
pyproj==2.4.2.post1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
Rtree==0.8.3
scipy==1.3.3
seaborn==0.9.0
Shapely==1.6.4.post2
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.2.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
urllib3==1.25.7
entrypoints==0.3
esy-osm-pbf==0.0.1
esy-osm-shape==0.0.3
